requirements.txt
pandas
numpy
matplotlib
python-dotenv
pymongo[srv]==3.6
pymongo
certifi
scikit-learn
-e .
